[CI] Avoid using pyfakefs's fs fixture
This took roughly a work day to track down. This issue only presented
itself in certain conditions, based on how pytest-xdist split up the
tests being run, into workers (e.g. a certain number of tests, and a
certain value for -n). Originally the issue only showed up on GHA, but
after much experimenting and shelling into the GHA Runner and playing
around, I finally found that I was able to reproduce it locally as well
by forcing -n to 4. But only on a branch that had more than a certain
number of tests so that the grouping worked out to trigger the issue.

The issue manifested itself with this, seemingly completely unrelated
error:

>       assert renderer, ".accepted_renderer not set on Response"
E       AssertionError: .accepted_renderer not set on Response

.tox/py311/lib/python3.11/site-packages/rest_framework/response.py:55: AssertionError

What is happening here is that the 'fs' fixture makes global changes to
mock out the filesystem, and ultimately the issue is that it doesn't
seem to clean up after itself (tear down) properly.

The exact details are a bit fuzzy to me still - and I've spent enough
time trying to track this down - but what seems to be happening is that
when it doesn't teardown properly, every other test in the same
pytest-xdist worker that is expecting an error response fails
spectacularly. In the end, Django tries to render an error template but
it can't even find that, because the mocked filesystem is still in play.

The seemingly easy solution here, which seems to work fine, is to stop
using the 'fs' fixture, and instead use a context manager to set up the
files we need for this particular test. The context manager properly
cleans up after itself, and things end up being green again.

Signed-off-by: Rick Elrod <rick@elrod.me>