Equal created and modified datetime for new records
New instances of model classes inherited from `CommonModel` are
created with `created_on` and `modified_on` fields values that are not
equal. Typically the `modified_on` field of a new record is either
expected to be `None` or equal to `created_on` field value, so that one
can determined if the record has been ever modified after creation.

This patch fixes function `now()` beeing called twice when saving a new
record.