Fix resource api signals connection
The `post_save` and `post_delete` signals intended to create
resources in DAB tables were connected in the `post_migrate` signal.

The `post_migrate` signal is not executed at the application startup,
therefore the signals were not connected.

This fixes `DoesNotExist` exception that occured in the
`create_resource` method of `Resource` model, when the resource that was
expected to be created by the `post_save` signal was queried:

    resource = cls.objects.get(object_id=content_object.pk, content_type=c_type)