Disconnect resource registry signals during migrations
This doesn't fix any particular immediate issue, just a nice to have
change. The `post_save` and `post_delete` signals, that are connected to
the `update_resource` and `delete_resource` receivers respectively,
are triggered during migrations for the following objects:

1. Migrations themselves
   (`django.db.migrations.recorder.MigrationRecorder.Migration.<locals>.Migration`).
2. Instances of fake models, dynamically created by Django for
   migrations (e.g. `__fake__.ContentType').

It's unlikely we want to run resource handlers for these kinds of
models, therefore disabling them in `pre_migrate` signal hook.