Rework how Activity Stream saves data
This is mostly just me playing around with an idea and seeing what
happens. This was inspired by #277 which documents that it's very hard
to deep-filter a JSONField.

The root of the problem is that JSONField is unstructured, so we don't
have type information up front about the values inside, and that -
with Django's Q-querying for JSONFields, not having that type
information means it's impossible to coerce the user's search
parameters (always a string) into the type we need for matching to
work in all the non-string cases.

The approach taken in this commit does things differently. Instead of
storing a JSONField with the changes, I introduce a FieldChange model,
which is essentially a glorified tuple of (stream_entry, operation,
field_name, old_value, new_value).

operation is necessary because it's how we know if fields were added
or removed from the model itself, versus being set explicitly null.

When we store an activity stream entry, we also store the associated
field changes. We always store the values as strings, so that the
filtering problem in #277 is never an issue, matching can always
work. In the new serializer, we reverse the string back into its
python value using Django's Field#to_python method. This means we
render JSON that has the right types for things like numbers and
booleans, but internally they're stored as strings, and can be
filtered with the querystring filters as you'd expect.

Effectively this lets us do things like:

?changes__field_name__exact=email&changes__new_value=admin@stuff.invalid

or

?changes__field_name__exact=id&changes__new_value=2

and other cool queries like that, into the actual fields being
changed.

This is a completely different approach from the JSONField approach,
but there's really no reason that we *had* to use a JSONField here. So
this attempts to explore what happens when we don't.

Signed-off-by: Rick Elrod <rick@elrod.me>