[ModifiableModel] Pass update_fields back to super (#299)
This came up when debugging ansible#286 which was supposed to be an easy fix
but led me down quite a rabbit hole.

For as long as modified_by has existed (from what I can tell in the
git history), we never obeyed update_fields in save(). When we split
it out into ModifiableModel, the bug was kept in tact.

What ends up happening here is that on login, we'd act as if we're
going to update the modified_by field, and we'd get that in
update_fields, but then we'd just eat update_fields and never pass it
back into the supermethod.

This led to weird things like activity stream thinking that the value
was getting updated when a user logged in (because contractually, it
was *supposed* to be getting updated), but since we forgot
update_fields, it didn't *actually* get updated.

So you'd see an activity stream entry with the change, but check in
the database and see that modified_by never actually updated on login.

This ensures we pass back in update_fields if we were given it in the
first place, and that we only update it when it makes sense (if it's
actually an update, i.e. self.pk exists).

I also add a test showing that if you update modified_by and set
update_fields to include it, the forced value you set it to gets saved
as expected.

Finally, this does *not* fix ansible#286. So I added an XFAIL test for
that. This was found while *working* on ansible#286, but it's a different
bug.

Signed-off-by: Rick Elrod <rick@elrod.me>