Fix validation error for role definition PATCH (#293)
The issue described to me - 

PATCH a role definition with a new name. This was in eda-server, which
doesn't allow creating custom roles which are system-wide. The validator
erroneously gives a 400 Bad Request, because it incorrectly gets
`content_type` from the request data, thinking that the action is
changing the type of the RoleDefinition to None, making it a system-wide
role.

The fix is that we must get the "new" field values either from the data
or from the model.