Remove object_role from RBAC API, cleanup logging (#264)
This is a grab-bag of a couple of things, which I'd like to address in
DAB before feature branches merge.

### Permission Logging

Steps to reproduce are:
 - give user view / member permission to a team
 - log in as that user
 - navigate to the team detail page

Doing this, DAB RBAC permission class gave a whole lot of WARNING log
messages. This is because DRF re-calls the permission class for other
methods like PUT/PATCH/etc. even though the request is obviously just a
GET request. This is distracting, and would make users lose trust in the
log messages that tell _the exact permission_ that the user lacks, and
thus the reason for 403 errors.

With this change, I believe logs from the permission class should be
actually telling the details of a denied request, which is what I want.

### API Stabilization

Here is user_role_assignments

```json
{
    "count": 3,
    "next": null,
    "previous": null,
    "results": [
        {
            "id": 1,
            "url": "/api/v1/role_user_assignments/1/",
            "related": {
                "created_by": "/api/v1/users/1/",
                "role_definition": "/api/v1/role_definitions/1/",
                "user": "/api/v1/users/5/",
                "content_object": "/api/v1/organizations/1/"
            },
            "summary_fields": {
                "created_by": {
                    "id": 1,
                    "username": "_system",
                    "first_name": "",
                    "last_name": ""
                },
                "role_definition": {
                    "id": 1,
                    "name": "Organization admin permissions",
                    "description": "",
                    "managed": true
                },
                "user": {
                    "id": 5,
                    "username": "org_admin",
                    "first_name": "",
                    "last_name": ""
                },
                "content_object": {
                    "id": 1,
                    "name": "AWX_community"
                }
            },
            "created": "2024-04-04T19:00:15.143023Z",
            "created_by": 1,
            "content_type": "shared.organization",
            "object_id": "1",
            "role_definition": 1,
            "user": 5
        },
```

This removes "object_role", both from the fields _and_ from the
`summary_fields` which is an issue of the shared common serializer, ping
@john-westcott-iv @relrod

We don't want to surface this, because it is internal, and we want to
claim full flexibility to change the details of these models (and some
talks of this have already happened).

### Unordered Queryset

I want to address the `UnorderedObjectListWarning` warnings by changing
the Meta properties. This doesn't actually result in a schema change, so
I'm not worried about downstream effect from putting it in the 0001
migration, but neither AWX or eda-server branches have merged yet so it
wouldn't matter anyway.

### Related Names

Using test_app, I wanted to be able to link from external models (like
Team and User) to role assignments. If you dig in, you find out that you
can't do that without a `related_name`, and this requires a migration,
so that's another thing I want to slip in before the deadline.