Alter expected_log to allow for checking multiple levels of same logger
Before this change you could not do:
   with expect_log(logger, 'debug', 'debug message'):
       with expect_log(logger, 'info', 'info message'):

Because the fixture was mocking logger and then requesting the severity.
So the first call would mock logger and the second call would remock logger.

This alters the code to mock to logger.severity allowing the scenario above to exist.

Note: This would not address attempting to test multiple messages from the same logging level