[AS] Address incorrect FK refs in summary_fields
Followup to #313. Don't use the current FK, use the one that it was
set to when the AS entry was logged.

This ended up triggering a lot more queries because there's really no
way to cache this effectively. So that led to reverting back to not
showing the extra summary_fields and related on list views.

Additionally, clean up a lot of redundant, verbose code, and make
things easier to read.

Signed-off-by: Rick Elrod <rick@elrod.me>