Fix mysterious schema generating error (#347)
I feel really bad that I don't have a test for this, but the test_app
docs works fine. I was able to replicate the issue with eda-server

http://127.0.0.1:8000/api/eda/v1/openapi.json

And I was able to confirm this fixes it (no 500 error). But I don't know
of any way we can test that here. In any case, the bug is obvious.