Allow use of role-tracking stuff in reverse relationship (#339)
I kind of remember intentionally not doing this before. At the time of
development, I was probably thinking I would return to it, and then I
thought we would remove this entire functionality, see
https://github.com/ansible/django-ansible-base/issues/159

But now we're using it, and things were broken. Looked at it, like hmmm,
is this because `reverse=True` in the signals? And yep, it was.

Clarifying: the test added here is the _first_ test to test this
synchronization (role to relationship) for a reverse relationship.