Partially migrate additional user editing conditions from AWX (#360)
Link the legacy code here:

https://github.com/ansible/awx/blob/918d5b3565e5997bdbd57117b92852ba13bb55d6/awx/main/access.py#L713-L726

When you look at that, first note that there is a condition for "orphan"
users --> this means users who are not in any organizations.

The need for this is clear from the test. If a user is not in _any_
organization, then we can't allow _any other user_ to manage that user.
That's why we need an extra condition in this case. That's not up for
debate.

Here, I'm using `is_superuser`, because I'm reluctant to copy over all
the rules from AWX to DAB. This one, in particular, seems dubious. The
reason why the AWX criteria had extra stuff was so that organization
admins could POST to `/api/v2/users/` and then after that add the user
as a member of their organization. I would like to punt this over to
require an API solution, so that a user can be given an organization on
creation @john-westcott-iv @slemrmartin