Allow configuring permissions needed to related objects (#346)
There was a sentiment that requiring "change" permission to related
objects might be too much. I expected this, but was trying to finish
development fast earlier.

Normally I would expect an app to rely on "use" permission, like AWX for
`use_role` of credentials and other things. But that requires
re-configuring permission structure and another permission for each
model. I'm not ready to take that step now, and without that your
choices are "change" or "view". If required is downgraded to "view",
that's perfectly coherent. It's a policy question that I have no
specific opinion on.

In short, this will allow apps using the DAB RBAC API utilities to
configure what level of related permission they want to check. This
comes in a precedence list defined in setting. If nothing in the
preference list applies to the target model, then we won't check
permissions for that item.