Fix bug detecting running container that needs migration (#364)
I've been hitting this a lot lately, and the prior solution just wasn't
working --> it went to error state when given a perfectly good running
(but un-migrated) database.

```
python manage.py migrate --check; migrate_needed=$?
```

This was supposed to capture the exit code, but it didn't. In the first
command, it just went straight to the trap. This seems to fix it... as
devious and ugly as it is.

---------

Co-authored-by: Rick Elrod <rick@elrod.me>