Move the rest_filters reserved names to a setting (#384)
My 2 cents here - this code was really just copied over from AWX. There
is a process that code needs to go through in order to generalize it so
that it's just abstract, for use via a library. This never happened for
at least this part of rest_filters. Since it came up, I wanted to go
ahead and get it done.