Fix 500 on unintended endpoints and further refactor association router (#370)
Something drove me crazy after testing with my last patch - which is
that certain serializers (with the same name) were created twice. This
is logical if you think about what's being done. This PR adds a registry
to fix the problem.

It also trades a few things in code organization. It gets rid of the
"factory" method, and adds a new utility method for getting the parent
class, which has to follow a certain format. Then the registry
management becomes the main job of `get_serializer_class`

Recapping of major changes being done here:

- Serializers for association & disassociation are refactored so that
they don't do "rbac" filtering by their own logic. This change gets rid
of the user-specialized association serializer using `visible_users` and
the general `access_qs` call in favor of calls to `filter_queryset`.
- As a subpoint of the above point, `get_serializer_class` no longer
builds querysets (for `instances`). This changes it to use methods from
the view _instance_ inside of the serializer context. This has to happen
on the serializer `__init__`.
- Disassociation validation that the object is already attached is
deleted, replaced by queryset logic, which now uses the same
`get_queryset` as the listing does, which only includes attached objects
to begin with.
- New parent class for related model viewsets is added which is inserted
and _lower_ precedence than the passed-in class. This means that methods
defined here can be overwritten in order to customize behavior. These
are listed in docs too.
- When synthesizing the new association viewset class, all methods
_other than list_ are disabled by a metaclass. This avoids the duplicate
registration I complained above above.
- A serializer registry is added so that we don't create the same-name
class multiple times, which confuses OpenAPI.