Fix dir() giving wrong answer in autogenerated viewsets (#410)
I only just merged
https://github.com/ansible/django-ansible-base/pull/370 but this fixes a
problem with it.

This was encountered using drf-spectacular decorators to resolve very
boutique eda-server issues, which could result in the traceback:

```
  File "/home/alancoding/venvs/eda/lib64/python3.11/site-packages/drf_spectacular/drainage.py", line 168, in get_view_method_names
    return [
           ^
  File "/home/alancoding/venvs/eda/lib64/python3.11/site-packages/drf_spectacular/drainage.py", line 169, in <listcomp>
    item for item in dir(view) if callable(getattr(view, item)) and (
                                           ^^^^^^^^^^^^^^^^^^^
  File "/home/alancoding/repos/awx/testing/django-ansible-base/ansible_base/lib/routers/association_resource_router.py", line 231, in attribute_raiser
    raise AttributeError
AttributeError
```

This traceback is a problem even when you have no other context.
You can see `item for item in dir(view)`. Then one of those items gives
an `AttributeError`. I mean, the `dir` method is pretty much, by
definition, the set of attributes that won't give you an attribute
error.

Non-obvious, but since all of this is done inspecting the view _class_,
similar to the AttributeError, all the fixes have to be done on the
metacalss.