Use dynamic docstrings for sublists (#419)
Previously the `associate` and `disassociate` docstring was getting
rendered in the "description" of OPTIONS and in the browsable API. It
wasn't _the worst_ in terms of documentation, but it was
awkward and out-of-place.

You'll see this in the code, but if it's read-only, and only shows list,
the POST options won't be included. This doesn't give specific
docstrings for each endpoint, because I think the bigger picture is more
important for usability here.