Identify specific actions from non-viewset views for resource registry (#446)
See https://github.com/ansible/django-ansible-base/pull/400, we found a
case of something sending "service-metadata" inside of the
"resource_api_actions", so while I removed that because it aggressively
broke AWX checks, something else was depending on what it did. So the
solution is to use a name other than `action` obviously.

This is a speculative fix, I don't know how I can test it without a
great deal over work.