Correct the content type as necessary for older systems.
In older releases of DAB, the resourcetype table could have been populated with
a different contenty type. For example, in galaxy it would have used the group
model for shared.team because it didn't yet have a Team model. This patch attempts
to catch the integrtiy error and then fix the relationships.

No-Issue

Signed-off-by: James Tanner <tanner.jc@gmail.com>