Fix queryset of RoleDefinition-assignment list with new utils (#425)
Fixes https://github.com/ansible/django-ansible-base/issues/304

AAP-25895

A lot of changes were recently made to the contract for `related_views`
in the router. The docs are at:


https://github.com/ansible/django-ansible-base/blob/devel/docs/lib/routers.md#customization

So by using `filter_queryset` a lot of things will "just work", but the
views have not really attempted to adopt these abilities until now. This
sublist view is a good example where we obviously have to filter the
list.

---------

Co-authored-by: jctanner <tanner.jc@gmail.com>