Make ansible_id match the JWT data if creating new user (#512)
I thought I fixed the issue
where the "admin" user had a resource conflict. But then it was still a
conflict.

It was a conflict created by an entirely new mechanism. The JWT consumer
was creating users from the JWT data that had a different ansible_id.

NOTE: We seem to have another problem where the `service_id` is not set
for absolutely anything created in this file.

AAP-26659