[Low Prio] Add a sqlite3 check (#517)
I got 3 failures running this locally, and ability to address those
failures is the key thing I'd like to get out of this. I would like to
position us to be multi-vendor regarding the database as much as we can
reasonably do with minimal effort, and this is a nice start.