Reverse syncing from services to a resource server
In the interest of trying to prevent as many failure scenarios as
possible (though this is NOT failproof as @AlanCoding points out below
in the comments), we wrap the REST call and the local commit in a
transaction (unless a transaction is already present).

For modules which sync (users, orgs, teams), their `save()` method is
monkeypatched to wrap the whole save in a transaction, if necessary.

A pre_delete and post_save signal is used for doing the actual sync.

In order to prevent sync loops, we sneak an internal attribute on
instances that we create/fetch on requests from a resource server. In
the syncing function, we look for this attribute and abort syncing if
it's there and `True`.

Unit tests (somewhat gnarly ones) are provided. They attempt to ensure
that we're calling the right REST client methods when we need to,
bailing out early when we need to, rolling back when we need to, and
committing when we need to. Coverage should be pretty darn close to
100%.

A new setting, `DISABLE_RESOURCE_SERVER_SYNC` works to disable the
reverse syncing. We use it in `test_app` to avoid syncing on all but
the specific reverse syncing tests (the ones written in this PR).

Signed-off-by: Rick Elrod <rick@elrod.me>