Make stop-postgres target remove the container as before (#549)
With a recent merge of
https://github.com/ansible/django-ansible-base/pull/515 we changed
behavior which was relied on in a few places for dev environment stuff -
the postgres container won't auto-delete itself when stopped.

You can't get this behavior exactly with docker-compose, but this seems
close enough

This changes the behavior back to how it was before, in that `make
stop-postgres` will stop and remove the container. This makes the
bootstrap script a little less fragile.