Allow removing permissions even when they are invalid (#559)
Discovered in testing in https://github.com/ansible/galaxy_ng/pull/2202

The reason for fixing it in DAB is that we can't honestly fix this issue in
the signals there. If we caught the exception, then the permission
wouldn't be removed. And that's wrong, as fail-deadly kind of situation.

Removing permissions should never give an error in a situation where we
are _able_ to remove the permission. Because this would _prevent_ the
app from fixing validation violations, which is the opposite of what we
want.