Fix bug where is_registered did not consider app_label (#560)
Troubleshooting https://github.com/jctanner/galaxy_ng/pull/13 I ran into
a new type of problem.

It created 2 permissions for `view_containernamespace`. This is because
there were 2 models:
 - galaxy_ng.app.models.container.ContainerNamespace
 - pulp_container.app.models.ContainerNamespace

Diving in, the logic in `create_dab_permissions` loops over all models
in an app. This is ran for every app. In this case, the registered app
is the concrete model of `pulp_container.app.models.ContainerNamespace`.
However, when creating permissions for the "galaxy" app, it asks
`permission_registry.is_registered(klass)`, and it returns `True` for
`galaxy_ng.app.models.container.ContainerNamespace`. It did this,
because the "model_name" meta attribute matched. Because of that, it
creates 2 permissions with the same `codename` which violates later
assumptions.

Co-authored-by: jctanner <tanner.jc@gmail.com>