Give 400 respones when roles violate RBAC validators (#563)
Reported in https://issues.redhat.com/browse/AAH-3350

This really has to be fixed in the role creation logic (you shouldn't be
able to create such roles through the API) so that the roles in the
database don't violate the validation rules. But if you have invalid
roles around, it gives 500 errors, as reported. This changes those to
400 errors.