Add some nuance to the on-off logic for RESOURCE_SERVER (#603)
Due to dynaconf implementation details, I discovered that eda-server may
set `{"URL": None}` for the `RESOURCE_SERVER` setting. This is totally
fine with me, and should be the same as not defining that setting. But
it requires this code change.

Fixes https://github.com/ansible/django-ansible-base/issues/602

This also makes the new auth class a no-op when the resource server
is not defined, reusing that same logic.