Add option to automatically move users to a new authenticator (#601)
This adds a new field on the authenticator model called
`auto_migrate_users_to` which accepts a foreign key to another
authenticator. When this field is configured we will do the following:

- On login, collect any authenticator users from authenticators
configured to migrate to the currently in use authenticator
- For each of the discovered authenticators, call the
`move_authenticator_user_to` method on the authenticator plugin. This
lets the authenticator plugin know that the user is being moved and
gives it a chance to perform any cleanup before that happens.