AAP-50805 Ignore role content types not known to system in sync (#811)
The bug here is that AWX `periodic_resource_sync` will result in an
error.

It seems there was an attempt to log an error and continue on when
getting a `ValidationError`, it seems this only targeted the Django
validation error exception, leaving the more common DRF ValidationError
uncaught. This means that the issue broke the resource sync, globally.
That means the periodic sync was effectively broken by the referenced
bug.

We could change that, but we wanted to have a targeted and more
intentional special-case just for roles, and that is what this
implements.

If the field `RoleDefinition.content_type` references a type that does
not exist locally, we don't bother in the _periodic fallback_ sync and
just skip over that record. This is very sensible behavior, and it will
be reported as a no-op, as one should expect.